fixed possible NPE
git-svn-id: https://svn.codehaus.org/mule/trunk@1091 bf997673-6b11-0410-b953-e057580c5b09

+++ mule/src/java/org/mule/util/Utility.java
@@ -318,14 +318,16 @@ public class Utility
             return true;
         }
         File[] files = dir.listFiles();
-        for (int i = 0; i < files.length; i++) {
-            if (files[i].isDirectory()) {
-                if (!deleteTree(files[i])) {
-                    return false;
-                }
-            } else {
-                if (!files[i].delete()) {
-                    return false;
+        if(files!=null) {
+            for (int i = 0; i < files.length; i++) {
+                if (files[i].isDirectory()) {
+                    if (!deleteTree(files[i])) {
+                        return false;
+                    }
+                } else {
+                    if (!files[i].delete()) {
+                        return false;
+                    }
                 }
             }
         }
